separate pitch & upperLeft state
