+++ src/states/pitchState.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const pitchState = atom<number>({
+  key: "pitchState",
+  default: 0,
+});
